+++ src/server/game/Loot/Loot.cpp
@@ -193,7 +193,7 @@ void Loot::AddItem(LootStoreItem const& item, bool canBePersonal)
                         {
                             VirtualModifier modifier = VirtualModifier();
 
-                            uint32 dungeonLevel = uint32(member->GetMap()->GetDungeonLevel());
+                            uint32 dungeonLevel = member->GetMap()->GetDungeonLevel();
                             uint32 playerLevel = uint32(member->GetCappedItemLevel());
 
                             if (const InstanceTemplate* inst = sObjectMgr->GetInstanceTemplate(member->GetMapId()))
@@ -272,7 +272,7 @@ void Loot::AddItem(LootStoreItem const& item, bool canBePersonal)
                 // if player is in dungeon, override avg level and apply vLvl mod
                 if (const InstanceTemplate* inst = sObjectMgr->GetInstanceTemplate(player->GetMapId()))
                 {
-                    uint32 cappedDungeonLevel = uint32(player->GetMap()->GetCappedDungeonLevel(inst->softcapMod));
+                    uint32 cappedDungeonLevel = player->GetMap()->GetCappedDungeonLevel(inst->softcapMod);
 
                     modifier.dungeonLevel = cappedDungeonLevel;
                     modifier.vLvlMod = inst->vLvlMod;
